requirements.txt
Flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
supabase==2.3.5
finnhub-python==2.4.19
yfinance==0.2.36
pandas==1.5.3
numpy==1.24.3
plotly==5.18.0
gunicorn==21.2.0 